current webpage
this has no content but a working navbar that still needs alot of work to change the placement and add the logo on the right, same as the photo slider, i cant change the pictures sizes and get them working from home. You can make any changes you want or scrap the whole thing if you think its crap mate.